refactor controllers and create services
